Get Configured Asset Sources

Get Configured Asset Sources

Usage

get_configured_asset_paths(assets_config, verbose = TRUE)

Arguments

assets_config

List containing asset configuration settings as part of the `napistu_config`. If this is an empty list, the `test_pathway` bundled with `napistu-r` will be used.

assets_dir

relative to this directory.

sbml_dfs

Path to a pickled `SBML_dfs` object generated by the Napistu Python package. `SBML_df`s are self-contained relational representations of a pathway.

napistu_graph

Path to a pickled `NapistuGraph` object generated by the Napistu Python package. `NapistuGraph`s are a subclass of `igraph.Graph`.

species_identifiers

Path to a .tsv containing the systematic identifiers for all of the molecular species in the `SBML_dfs`.

precomputed_distances

Optional, path to a .parquet file containing pairwise distances between molecular species. Defining this can speed up the calculation of shortest paths and neighborhoods.

verbose

Logical scalar; if TRUE (default), prints informative messages about the current operation progress. If FALSE, runs silently.

Value

Named list of asset file paths for configured assets