Setup Napistu List

Load a config specifying a Python environment and paths to Napistu assets as a list providing defaults as fall-backs.

setup_napistu_list(
  napistu_config,
  napistu_list_object = NAPISTU_CONSTANTS$NAPISTU_LIST_OBJECT,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

napistu_config

A napistu configuration object of class 'napistu_config' containing all necessary configuration settings for the loading required assets and configuring and/or using an exiting appropriate Python environment

napistu_list_object

The name that setup_napistu_env() will be assigned to. If this object already exists then setup will be skipped unless `overwrite` is TRUE

overwrite

Overwrite existing caches

verbose

Logical scalar; if TRUE (default), prints informative messages about the current operation progress. If FALSE, runs silently.

Value

None; output is assigned to `napistu_list_object`

Examples

if (FALSE) { # \dontrun{
napistu_config <- create_napistu_config()
setup_napistu_list(napistu_config)

# Using config file
setup_napistu_list("my_config.yml")
} # }